backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
PyJWT==2.8.0
websockets==12.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pyvisa==1.14.1
numpy==1.26.2
matplotlib==3.8.2
pandas==2.1.4
fpdf2==2.7.6
aiofiles==23.2.1